move VS daily PDF find logic and strip function to VS common module
