+++ scrapers/scrape_vs_common.py
@@ -33,3 +33,26 @@ def get_vs_weekly_general_data(pdf):
     start_date = end_date - datetime.timedelta(days=7)
     year = start_date.year
     return week, year
+
+
+def get_vs_all_daily_pdf_url():
+    base_url = 'https://www.vs.ch'
+    url = f'{base_url}/web/coronavirus/statistiques'
+    content = sc.download(url, silent=True)
+    soup = BeautifulSoup(content, 'html.parser')
+    links = soup.find_all('a', string=re.compile(r'20\d{2}.*Sit Epid.*'))
+    pdf_urls = []
+    for link in links:
+        pdf_urls.append(f"{base_url}{link.get('href')}")
+    return pdf_urls
+
+
+def get_vs_daily_pdf_url():
+    urls = get_vs_all_daily_pdf_url()
+    return urls[0]
+
+
+def strip_value(value):
+    if value:
+        return re.sub(r'[^0-9]', '', value)
+    return None
